Extracted SessionWriter from Session. Extracted ModWriter from Mod.
